pyproject.toml
[tool.poetry]
name = "langgraph-bedrock-converse"
version = "0.1.0"
description = ""
authors = ["Dan <danseaton@kubrickgroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
langchain = "^0.2.6"
langchain-aws = "^0.1.9"
boto3 = "^1.34.140"
python-dotenv = "^1.0.1"
langgraph = "^0.1.5"
httpx = "^0.27.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
